Move structs to own file
